+++ pkg/studyengine/actions_test.go
@@ -26,7 +26,7 @@ func TestActions(t *testing.T) {
 		action := types.Expression{
 			Name: "WRONG",
 		}
-		_, err := ActionEval(action, participantState, event)
+		_, err := ActionEval(action, participantState, event, nil)
 		if err == nil {
 			t.Error("should return an error")
 		}
@@ -52,7 +52,7 @@ func TestActions(t *testing.T) {
 				{DType: "exp", Exp: &action2},
 			},
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -68,7 +68,7 @@ func TestActions(t *testing.T) {
 				{DType: "exp", Exp: &action3},
 			},
 		}
-		newState, err = ActionEval(action, participantState, event)
+		newState, err = ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -97,7 +97,7 @@ func TestActions(t *testing.T) {
 				{DType: "exp", Exp: &action2},
 			},
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -112,7 +112,7 @@ func TestActions(t *testing.T) {
 				{DType: "exp", Exp: &action2},
 			},
 		}
-		newState, err = ActionEval(action, participantState, event)
+		newState, err = ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -128,7 +128,7 @@ func TestActions(t *testing.T) {
 				{DType: "exp", Exp: &action3},
 			},
 		}
-		newState, err = ActionEval(action, participantState, event)
+		newState, err = ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -145,7 +145,7 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: "value"},
 			},
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -166,7 +166,7 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: "health"},
 			},
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -189,7 +189,7 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: types.ASSIGNED_SURVEY_CATEGORY_NORMAL},
 			},
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -217,11 +217,11 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: types.ASSIGNED_SURVEY_CATEGORY_NORMAL},
 			},
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
-		newState, err = ActionEval(action, newState, event)
+		newState, err = ActionEval(action, newState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -234,7 +234,7 @@ func TestActions(t *testing.T) {
 		action = types.Expression{
 			Name: "REMOVE_ALL_SURVEYS",
 		}
-		newState, err = ActionEval(action, newState, event)
+		newState, err = ActionEval(action, newState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -256,17 +256,17 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: types.ASSIGNED_SURVEY_CATEGORY_NORMAL},
 			},
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
 		action.Data[0].Str = "testSurveyKey2"
-		newState, err = ActionEval(action, newState, event)
+		newState, err = ActionEval(action, newState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
 		action.Data[0].Str = "testSurveyKey1"
-		newState, err = ActionEval(action, newState, event)
+		newState, err = ActionEval(action, newState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -283,7 +283,7 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: "last"},
 			},
 		}
-		newState, err = ActionEval(action, newState, event)
+		newState, err = ActionEval(action, newState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -309,17 +309,17 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: types.ASSIGNED_SURVEY_CATEGORY_NORMAL},
 			},
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
 		action.Data[0].Str = "testSurveyKey2"
-		newState, err = ActionEval(action, newState, event)
+		newState, err = ActionEval(action, newState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
 		action.Data[0].Str = "testSurveyKey1"
-		newState, err = ActionEval(action, newState, event)
+		newState, err = ActionEval(action, newState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -336,7 +336,7 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: "first"},
 			},
 		}
-		newState, err = ActionEval(action, newState, event)
+		newState, err = ActionEval(action, newState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -361,16 +361,16 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: types.ASSIGNED_SURVEY_CATEGORY_NORMAL},
 			},
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
-		newState, err = ActionEval(action, newState, event)
+		newState, err = ActionEval(action, newState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
 		action.Data[0].Str = "testSurveyKey2"
-		newState, err = ActionEval(action, newState, event)
+		newState, err = ActionEval(action, newState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -386,7 +386,7 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: "testSurveyKey1"},
 			},
 		}
-		newState, err = ActionEval(action, newState, event)
+		newState, err = ActionEval(action, newState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -427,7 +427,7 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: "test.2.2"},
 			},
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -443,7 +443,7 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: "test.2.1"},
 			},
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -456,7 +456,7 @@ func TestActions(t *testing.T) {
 		action := types.Expression{
 			Name: "REMOVE_ALL_REPORTS",
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -473,7 +473,7 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: "first"},
 			},
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -494,7 +494,7 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: "last"},
 			},
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
@@ -514,7 +514,7 @@ func TestActions(t *testing.T) {
 				{DType: "str", Str: "test.1"},
 			},
 		}
-		newState, err := ActionEval(action, participantState, event)
+		newState, err := ActionEval(action, participantState, event, nil)
 		if err != nil {
 			t.Errorf("unexpected error: %s", err.Error())
 		}