+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - `IF`: improved method for control flows with if-else logic
     - `DO`: to perform a list of actions
 - Some documentation about study actions [here](docs/studyActions.md).
+- New study expression: `checkConditionForOldResponses`
 
 ## [v0.14.1]